remove CreditsReceiver from society to society_manage
