+++ society_manage/models.py
@@ -1,3 +1,12 @@
 from django.db import models
+from society.models import Society, Student
 
-# Create your models here.
+
+class CreditReceivers(models.Model):
+    society = models.ForeignKey(Society, on_delete=models.DO_NOTHING)
+    receivers = models.ForeignKey(Student, on_delete=models.DO_NOTHING)
+    year = models.PositiveSmallIntegerField()
+    semester = models.PositiveSmallIntegerField()
+
+    def __str__(self):
+        return str(self.society)
